Added updating comment by anonymous user

+++ src/main/java/com/course/leverxproject/service/comment/CommentServiceImpl.java
@@ -2,6 +2,7 @@ package com.course.leverxproject.service.comment;
 
 import com.course.leverxproject.dto.comment.CommentCreateRequestDTO;
 import com.course.leverxproject.dto.comment.CommentResponseDTO;
+import com.course.leverxproject.dto.comment.CommentResponseWithTokenDTO;
 import com.course.leverxproject.dto.comment.CommentUpdateRequestDTO;
 import com.course.leverxproject.entity.Comment;
 import com.course.leverxproject.entity.User;
@@ -10,16 +11,23 @@ import com.course.leverxproject.exception.user.SellerNotFoundException;
 import com.course.leverxproject.repository.CommentRepository;
 import com.course.leverxproject.repository.UserRepository;
 import com.course.leverxproject.service.auth.AuthService;
+import com.course.leverxproject.service.auth.MyUserDetails;
+import com.course.leverxproject.service.jwt.JwtService;
 import com.course.leverxproject.service.user.UserService;
+import com.course.leverxproject.utils.SecurityUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -33,21 +41,33 @@ public class CommentServiceImpl implements CommentService {
 
     private final UserService userService;
 
-    public CommentServiceImpl(CommentRepository commentRepository, UserRepository userRepository, AuthService authService, UserService userService) {
+    private final JwtService jwtService;
+
+    public CommentServiceImpl(CommentRepository commentRepository, UserRepository userRepository, AuthService authService, UserService userService, JwtService jwtService) {
         this.commentRepository = commentRepository;
         this.userRepository = userRepository;
 
         this.authService = authService;
         this.userService = userService;
+        this.jwtService = jwtService;
     }
 
 
     @Override
-    public CommentResponseDTO createComment(int userId, CommentCreateRequestDTO commentDTO) {
-        //TODO Add checking if seller has role Seller.
+    public CommentResponseWithTokenDTO createComment(int userId, CommentCreateRequestDTO commentDTO) {
+
         User seller = userRepository.findById(userId).orElseThrow(() -> new SellerNotFoundException("Seller with id " + userId + " not found"));
-        //TODO Consider what to do to set this user in session (Dont create users every time in one session)
-        User author = authService.createAnonymous();
+        if (seller.getRoles().stream().noneMatch(role -> role.getName().equals("ROLE_SELLER"))) {
+            throw new SellerNotFoundException("Cannot find seller with " + userId + "id");
+        }
+        Optional<MyUserDetails> currentUser = SecurityUtils.getCurrentUser();
+        User author;
+
+        if (currentUser.isPresent()) {
+            author = currentUser.get().getUser();
+        } else {
+            author = authService.createAnonymous();
+        }
 
         Comment comment = new Comment(
                 commentDTO.message(),
@@ -59,7 +79,10 @@ public class CommentServiceImpl implements CommentService {
         );
 
         commentRepository.save(comment);
-        return new CommentResponseDTO(
+        String token = jwtService.generateToken(String.valueOf(author.getId()), "ANONYMOUS");
+        Authentication auth = jwtService.getAuthentication(token);
+        SecurityContextHolder.getContext().setAuthentication(auth);
+        return new CommentResponseWithTokenDTO(new CommentResponseDTO(
                 comment.getId(),
                 comment.getMessage(),
                 comment.getRate(),
@@ -68,35 +91,39 @@ public class CommentServiceImpl implements CommentService {
                 comment.getSeller().getFirstName() + " " + comment.getSeller().getLastName(),
                 comment.getCreatedAt(),
                 comment.getApproved()
+        ), token
         );
     }
 
     @Override
     public CommentResponseDTO updateComment(int commentId, CommentUpdateRequestDTO commentDTO) {
-        //TODO Add validation (only author of comment cant comment)
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException("Comment with id " + commentId + " not found."));
+
+        int currentUserId = SecurityUtils.getCurrentUser().map(MyUserDetails::getId).orElseThrow(() -> new AccessDeniedException("User not authenticated"));
+
+
+        if (comment.getAuthor().getId() != currentUserId) {
+            throw new AccessDeniedException("User not authorized to update this comment.");
+
+        }
+
         comment.setMessage(commentDTO.message());
         comment.setRate(commentDTO.rate());
         commentRepository.save(comment);
 
-        return new CommentResponseDTO(
-                comment.getId(),
-                comment.getMessage(),
-                comment.getRate(),
-
-                comment.getAuthor().getFirstName(),
-
-                comment.getSeller().getId(),
-                comment.getSeller().getFirstName() + " " + comment.getSeller().getLastName(),
-                comment.getCreatedAt(),
-                comment.getApproved()
-        );
+        return commentToCommentResponseDTO(comment);
     }
 
     @Override
     public void deleteComment(int commentId) {
-        //TODO Check if user has permission to do that.
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException("Comment with id " + commentId + " not found."));
+        int currentUserId = SecurityUtils.getCurrentUser()
+                .map(MyUserDetails::getId)
+                .orElseThrow(() -> new AccessDeniedException("User not authenticated."));
+        if (comment.getAuthor().getId() != currentUserId) {
+            throw new AccessDeniedException("User not authorized to delete comment.");
+        }
+
         commentRepository.delete(comment);
     }
 
@@ -104,18 +131,9 @@ public class CommentServiceImpl implements CommentService {
     public CommentResponseDTO getComment(int commentId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException("Comment with id " + commentId + " not found."));
 
-        return new CommentResponseDTO(
-                comment.getId(),
-                comment.getMessage(),
-                comment.getRate(),
-                comment.getAuthor().getFirstName(),
-
-                comment.getSeller().getId(),
-                comment.getSeller().getFirstName() + " " + comment.getSeller().getLastName(),
-                comment.getCreatedAt(),
-                comment.getApproved()
-        );
+        return commentToCommentResponseDTO(comment);
     }
+
 
     @Override
     public List<CommentResponseDTO> getComments(int userId, int page, int size, String sortBy, String sortDir) {
@@ -124,18 +142,8 @@ public class CommentServiceImpl implements CommentService {
         Page<Comment> comments = commentRepository.findAllBySellerId(userId, pageable);
         return comments
                 .stream()
-                .map(comment -> new CommentResponseDTO(
-                        comment.getId(),
-                        comment.getMessage(),
-                        comment.getRate(),
-                        comment.getAuthor().getFirstName(),
-
-                        comment.getSeller().getId(),
-                        comment.getSeller().getFirstName() + " " + comment.getSeller().getLastName(),
-                        comment.getCreatedAt(),
-                        comment.getApproved()
-
-                )).toList();
+                .map(this::commentToCommentResponseDTO
+                ).toList();
     }
 
     @Override
@@ -146,5 +154,19 @@ public class CommentServiceImpl implements CommentService {
         commentRepository.save(comment);
         userService.updateAverage(comment.getSeller().getId());
 
+    }
+
+
+    private CommentResponseDTO commentToCommentResponseDTO(Comment comment) {
+        return new CommentResponseDTO(
+                comment.getId(),
+                comment.getMessage(),
+                comment.getRate(),
+                comment.getAuthor().getFirstName(),
+                comment.getSeller().getId(),
+                comment.getSeller().getFirstName() + " " + comment.getSeller().getLastName(),
+                comment.getCreatedAt(),
+                comment.getApproved()
+        );
     }
 }
